Move code that populates the device tree into a new class and module.
